Upbit_AutoTrader_MVP_v1/requirements.txt
python-telegram-bot>=21.9
pyupbit>=0.2.33
pandas>=2.2.2
numpy>=1.26.4
APScheduler>=3.10.4
python-dotenv>=1.0.1
pytz>=2024.1
